Fix syntax error when executing setup.py using python 3.4
There are SyntaxError messages during processing reader.py and
cwise_ops_test.py due to incompatible syntax of python 2 and 3.

According to 'PEP 3113 -- Removal of Tuple Parameter Unpacking',
a tuple parameter is deprecated from 3.0 version of python, so we
modify method _build_vocab to make it compatible with python 2 and
python 3.