Fix an issue on token overwriting in conll2tree.py of SyntaxNet
There is an issue of token overwriting in conll2tree of SyntaxNet. Because parsed trees are build using dictionary and dictionaries use a token as a key, if the same token appears more than once on the same dictionay, the previous key-value pair is overwritten by new one.

For example, if input is given as 'Two iced Americanos, two plain lattes and two vanilla lattes.', there are repeated words ('lattes') on the same level. Then the parsed tree is given by:

Americanos NNP ROOT
 +-- Two CD num
 +-- iced JJ amod
 +-- , , punct
 +-- lattes NNS conj
 |   +-- two CD num
 |   +-- vanilla NN amod
 +-- and CC cc
 +-- . . punct

Note that the nodes on the first 'lattes' are gone.
So to solve this issue, I suggest to include each token's identifier to key for items of dictionary, and the index of token in original input is good candidate because they are unique and we can map tokens of the parse tree into original sentence exactly.

As a result, the new output will be:

Americanos NNP ROOT 2
 +-- Two CD num 0
 +-- iced JJ amod 1
 +-- , , punct 3
 +-- lattes NNS conj 6
 |   +-- two CD num 4
 |   +-- plain JJ amod 5
 +-- and CC cc 7
 +-- lattes NNS conj 10
 |   +-- two CD num 8
 |   +-- vanilla NN amod 9
 +-- . . punct 11