Add an input Tensor to BeamParseReader to provide serialised sentence protos to parse
Add an input Tensor to BeamParseReader to provide serialised sentence protos to parse.
Documents may come from the input tensor (e.g. read by a DocumentSource op)
or read directly from the corpus as defined in the task context. The decision
of the approach should is explicit based on attributes during graph configuration.