checkpoint/data collection flexibility
Adding hook to command-line argument for data/checkpoint save frequency and maximum checkpoint to keep.
This is helpful in-case for get time to accuracy measurements. For this measurement we want to specify how often we need to save the checkpoint - may be for every epoch, which depends on the batch-size. And also we want to keep all the checkpoint, not only five, to get the accuracy at the end of complete run. Printing log at the checkpoint-step will help to calculate the elapsed time.