Automated rollback of changelist 160744852 END_PUBLIC
*** Reason for rollback ***

Break FCNN model.

Some of the outputs of fused batch norm are used to compute gradients during training, if during evaluation/inference, these outputs are invalid/uninitialized numbers (there may or may not be NaNs in it).

There are use cases to set the "is_training" parameter of batch norm to False during training, and this will trigger the errors (or produce NaNs) because the "invalid/uninitialized numbers" got used.

Will re-enable after have a fix for this. The solution is proposed in a similar issue here: https://github.com/tensorflow/tensorflow/issues/10857

*** Original change description ***

BEGIN_PUBLIC
Enable fused batch norm for inception and resnet models.

PiperOrigin-RevId: 161138589