Fix bug in dataset.py when batch_size==max_length.
Each batch contains sequences of length in [buckets_min, buckets_max), so that the longest instance in a batch will have length buckets_max - 1. Without this fix, instances with length == batch_size will result in buckets of zero elements, causing the dataset iterator to crash.

As a bonus this adds an extra instance per batch.