Update installation.md
I did not realize that the tensorflow/models repository had to be cloned and setup. It looks like other people also did not realize this: https://github.com/tensorflow/models/issues/2031#issuecomment-317999079

The readme files in models/ and models/research do not mention anything about this either. Therefore, I updated this repository to mention that these steps must be done.