Update installation.md
When I read these instructions, I thought that the models/research directory was something that was located in my tensorflow directory that pip created. I did not realize that this repository had to be cloned. It seems that other people were confused by this as well: https://github.com/tensorflow/models/issues/2253

I have updated the instructions to try to indicate that the models directory is the root of this repository, and not part of pip's TensorFlow installation.