Fix training crash on visualizing Unicode
When the font `arial.ttf` is not found it loads the default font and crashes whenever it finds Unicode characters on both Python 2 and 3. A warning should be shown first before filtering all non-ASCII characters in case the user used all-Unicode-labels. 

The same error is shown when somebody tries `'百'.encode('latin-1')` in a Python interpreter.