Splitting the NCF embedding table when the accelerator is GPU.
Using a combined (bigger) embedding table followed by tensor slicing
results in an extra write and read operation, which can be completely
avoided if we maintain separate embedding tables. This results in a
significant performance improvement on the GPU.

As a sanity check, when I fixed the weights of the embedding to a
constant, and the weights were the same for all other layers before the
training, the value of the logits were the same after one step. After
the 1st step, the values will differ due to non-determinism in gradient
calculations.

For the following experiment, I used
  batch size = 99000
  synthetic data
  xla (on / off)

Without XLA
  Original: ~ 7.0 M examples/sec
  Modified: ~ 7.9 M examples/sec (12.8 % improvement)

With XLA:
  Original: ~ 9.2 M examples/sec
  Modified: ~ 9.5 M examples/sec (3.2 % improvement)

python ncf_keras_main.py \
    --model_dir /tmp/ \
    --data_dir data \
    --dataset ml-20m \
    --hooks --clean \
    --seed 3 \
    --train_epochs 3 \
    --batch_size 99000 \
    --eval_batch_size 99000 \
    --layers 256,256,128,64 \
    --num_factors 64 \
    --hr_threshold 0.635 \
    --ml_perf \
    --num_gpus 1 \
    --keras_use_ctl \
    --enable_xla \
    --use_synthetic_data