Add pylint wrapper extracted from TensorFlow CI. (#9793)
TensorFlow uses a subset of pylint functionality. Script will download the latest pylint rules from the main repo at runtime.

Requires `python3.8 -m pip install pylint==2.4.4`.
This will eventually be run on-PR in this repo, but first we'll have to fix the ~900 existing violations.
Script should be usable manually for now -- call it from within a subdirectory to avoid processing the entire repo.