Add output_stride parameter to quantize.md
The default output_stride in export_model.py is 8. The example in quantize.md is a model that expects an output_stride of 16. So if you try to use export_model.py without setting the output_stride, the resulting model will convert to tflite fine but with the wrong output_stride. This likely produces an invalid model. In my use case where I wanted to compile the model for an edgetpu, I suspect that this issue with the resulting model was causing the compiler to abort.