Return default strategy from get_distribution_strategy when given "off".
Before, it returned None. But almost every use of get_distribution_strategy() assumes an actual strategy is returned and crashes when None is returned. Returning the default strategy fixes these issues and is equivalent to using no strategy, as the default strategy is always in effect when no other strategy is used.

PiperOrigin-RevId: 380951055