Updated attention_ocr model to be compatible with TensorFlow 2.x.
Updated references to removed `tf.contrib` packages:
- `tf.contrib.slim` is now distributed as a separate PIP package tf-slim.
- `tf.contrib.legacy_seq2seq` is no longer available. The classes need by this model have been copied here.
- `tf.contrib.lookup.index_to_string_table_from_tensor` has been replaced with `tf.lookup.StaticHashTable`.
- `tf.contrib.layers.one_hot_encoding` has been replaced with `tf.one_hot`.

Added the FSNS list of URLs which used to be under `research/street` (that model was removed by PR #8934).
This update makes the `research/attention_ocr` model compatible with TensorFlow 2, but it is not TF2 native, since it uses compat.v1 layers. Checkpoints trained with TF1 can still be loaded with the latest version, but results my be slightly different due to implementation changes with respect to old contrib layers.

BUG: The `test_moving_variables_properly_loaded_from_a_checkpoint` in `demo_inference.py` no longer passes (probably to internal changes in SLIM). We need retrain the model and provide an updated checkpoint that makes this test pass.